// sesame-endpoint/src/gen/sesame/type/netaddr_pb.ts
// @generated by protoc-gen-es v2.15.0 with parameter "target=ts"
// @generated from file sesame/type/netaddr.proto (package sesame.type, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file sesame/type/netaddr.proto.
 */
export const file_sesame_type_netaddr: GenFile = /*@__PURE__*/
  fileDesc("ChlzZXNhbWUvdHlwZS9uZXRhZGRyLnByb3RvEgtzZXNhbWUudHlwZSIrCgdOZXRBZGRyEg8KB25ldHdvcmsYASABKAkSDwoHYWRkcmVzcxgCIAEoCUIsWipnaXRodWIuY29tL2pvZXljdW1pbmVzL3Nlc2FtZS90eXBlL25ldGFkZHJiBnByb3RvMw");

/**
 * NetAddr represents a network end point address.
 *
 * The two fields network and address are conventionally strings that can be passed as the arguments to Go's net.Dial,
 * but the exact form and meaning of the strings is up to the implementation.
 *
 * See also Go's net.Addr interface.
 *
 * @generated from message sesame.type.NetAddr
 */
export type NetAddr = Message<"sesame.type.NetAddr"> & {
  /**
   * @generated from field: string network = 1;
   */
  network: string;

  /**
   * @generated from field: string address = 2;
   */
  address: string;
};

/**
 * Describes the message sesame.type.NetAddr.
 * Use `create(NetAddrSchema)` to create a new message.
 */
export const NetAddrSchema: GenMessage<NetAddr> = /*@__PURE__*/
  messageDesc(file_sesame_type_netaddr, 0);